IS-1676: Add final state flag to vurdering dto
