+++ src/main/kotlin/no/nav/syfo/aktivitetskrav/domain/AktivitetskravStatus.kt
@@ -1,0 +1,26 @@
+package no.nav.syfo.aktivitetskrav.domain
+
+import java.util.EnumSet
+
+enum class AktivitetskravStatus {
+    NY,
+    AVVENT,
+    UNNTAK,
+    OPPFYLT,
+    AUTOMATISK_OPPFYLT,
+    FORHANDSVARSEL,
+    STANS,
+    IKKE_OPPFYLT,
+    IKKE_AKTUELL,
+    LUKKET,
+}
+
+fun AktivitetskravStatus.isFinal() = this in finalStatuses
+
+private val finalStatuses = EnumSet.of(
+    AktivitetskravStatus.UNNTAK,
+    AktivitetskravStatus.OPPFYLT,
+    AktivitetskravStatus.IKKE_OPPFYLT,
+    AktivitetskravStatus.IKKE_AKTUELL,
+    AktivitetskravStatus.AUTOMATISK_OPPFYLT,
+)
